% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{bcrp_data}
\alias{bcrp_data}
\title{Consulta BCRP API}
\usage{
bcrp_data(series, fechaini, fechafin, labels = "default", labs_opcional = NULL)
}
\arguments{
\item{series}{Vector de series a consultar del API.}

\item{fechaini}{Periodo inicial de consulta}

\item{fechafin}{Periodo final de la consulta}

\item{labels}{Permite incluir etiquetado a los indicadores de c("default", "manual","no"). En la opción "manual", el usuario puede añadir los labels de las variables en argumento complementario labs_opcional. Cuando la opción "no" se selecciona, no se incluyen labels.}
}
\value{
Retorna una base de datos con la(s) serie(s) en el periodo escogido.
}
\description{
Permite obtener de forma rápida información del API del BCRP
#'
}
\examples{
bcrp_data(series = c("PN01728AM"), fechaini = "2004-2", fechafin = "2019-12") # Mensual
bcrp_data(series = c("PN02507AQ"), fechaini = "2019-1", fechafin = "2022-4") # Trimestral
bcrp_data(series = c("PM04983AA"), fechaini = "2007", fechafin = "2022") # Anual

}
