% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{diccionario_bcrp}
\alias{diccionario_bcrp}
\title{Diccionario de la API}
\usage{
diccionario_bcrp()
}
\value{
Base de datos de todas las series disponibles en el API del BCRP
}
\description{
Consulta de series de datos disponibles en el API del BCRP según fecha, nombre, conjunto de datos, entre otros
}
\examples{
diccionario_bcrp() # Retorno de toda la serie
}
